refactor: extract the composer's surfaces out of MessageComposer.vue
