+++ resources/js/composables/useComposerKeyboard.ts
@@ -1,0 +1,151 @@
+import type { ComposerEditMode } from '@/composables/useComposerEditMode';
+import type { ComposerField } from '@/composables/useComposerField';
+import type { ComposerFormat } from '@/composables/useComposerFormat';
+import type { ComposerMentions } from '@/composables/useComposerMentions';
+import type { ComposerSlashCommands } from '@/composables/useComposerSlashCommands';
+import { isComposerEditTrigger } from '@/lib/composerEdit';
+import type { Message } from '@/types';
+
+/**
+ * The composer's keydown model, in priority order: whichever autocomplete menu
+ * is open owns the arrows, Enter/Tab and Escape; then the format shortcuts;
+ * then Escape's fallbacks (leave edit mode, else drop the reply); then ArrowUp's
+ * "edit last message" recall; and finally Enter to send or save.
+ */
+export function useComposerKeyboard(options: {
+    field: ComposerField;
+    mentions: ComposerMentions;
+    slash: ComposerSlashCommands;
+    format: ComposerFormat;
+    editing: ComposerEditMode;
+    replyTarget: () => Message | null | undefined;
+    onCancelReply: () => void;
+    onSubmit: () => void;
+}): { onKeydown: (event: KeyboardEvent) => void } {
+    const { body, caretAtStart } = options.field;
+    const { mentions, slash, editing } = options;
+
+    function onKeydown(event: KeyboardEvent): void {
+        if (mentions.showMenu.value) {
+            if (event.key === 'ArrowDown') {
+                event.preventDefault();
+                mentions.moveActive(1);
+
+                return;
+            }
+
+            if (event.key === 'ArrowUp') {
+                event.preventDefault();
+                mentions.moveActive(-1);
+
+                return;
+            }
+
+            if (event.key === 'Enter' || event.key === 'Tab') {
+                event.preventDefault();
+                mentions.selectActive();
+
+                return;
+            }
+
+            if (event.key === 'Escape') {
+                event.preventDefault();
+                mentions.closeMenu();
+
+                return;
+            }
+        }
+
+        // The slash-command menu mirrors the mention menu's navigation. The two are
+        // mutually exclusive (a `/…` body never matches an `@query`), so this runs
+        // only when the slash menu is the open one.
+        if (slash.showSlashMenu.value) {
+            if (event.key === 'ArrowDown') {
+                event.preventDefault();
+                slash.slashMoveActive(1);
+
+                return;
+            }
+
+            if (event.key === 'ArrowUp') {
+                event.preventDefault();
+                slash.slashMoveActive(-1);
+
+                return;
+            }
+
+            if (event.key === 'Enter' || event.key === 'Tab') {
+                event.preventDefault();
+                slash.selectSlashActive();
+
+                return;
+            }
+
+            if (event.key === 'Escape') {
+                event.preventDefault();
+                slash.closeSlashMenu();
+
+                return;
+            }
+        }
+
+        // Format shortcuts wrap the selection. Placed after the mention menu's key
+        // handling so its arrow/Enter/Escape keep priority while it is open; the
+        // chosen keys (B/I/E, ⇧X) never collide with it or Enter-to-send.
+        if (options.format.tryFormatShortcut(event)) {
+            return;
+        }
+
+        // With the mention menu closed, Escape leaves edit mode (restoring the empty
+        // composer) or, failing that, dismisses the active reply context.
+        if (event.key === 'Escape' && editing.editingMessage.value) {
+            event.preventDefault();
+            editing.exitEditMode();
+
+            return;
+        }
+
+        if (event.key === 'Escape' && options.replyTarget()) {
+            event.preventDefault();
+            options.onCancelReply();
+
+            return;
+        }
+
+        // ArrowUp on an empty composer recalls the viewer's last editable message
+        // into edit mode ("↑ to edit last message"). The gate keeps it clear of the
+        // mention menu, `⌥↑` channel nav, and an in-progress reply.
+        if (
+            isComposerEditTrigger({
+                key: event.key,
+                altKey: event.altKey,
+                ctrlKey: event.ctrlKey,
+                metaKey: event.metaKey,
+                shiftKey: event.shiftKey,
+                menuOpen: mentions.showMenu.value || slash.showSlashMenu.value,
+                editing: editing.editingMessage.value !== null,
+                hasReplyTarget: options.replyTarget() != null,
+                isEmpty: body.value.trim() === '',
+                caretAtStart: caretAtStart(),
+            })
+        ) {
+            if (editing.tryEditLastMessage()) {
+                event.preventDefault();
+            }
+
+            return;
+        }
+
+        if (event.key === 'Enter' && !event.shiftKey) {
+            event.preventDefault();
+
+            if (editing.editingMessage.value) {
+                editing.saveEdit();
+            } else {
+                options.onSubmit();
+            }
+        }
+    }
+
+    return { onKeydown };
+}
